+++ packages/terra-form-textarea/src/terra-dev-site/test/form-textarea/PopulatedTextarea.test.jsx
@@ -6,6 +6,7 @@ const textarea = () => (
     name="job_experience"
     defaultValue=""
     required
+    ariaLabel="label"
   />
 );