+++ packages/terra-form-textarea/src/terra-dev-site/doc/form-textarea/Textarea.1.doc.jsx
@@ -19,6 +19,8 @@ import InvalidExample from '../example/Invalid';
 import InvalidExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/Invalid';
 import DisabledExample from '../example/Disabled';
 import DisabledExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/Disabled';
+import TextareaExample from '../example/Textarea';
+import TextareaExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/Textarea';
 
 const DocPage = () => (
   <DocTemplate
@@ -28,6 +30,10 @@ const DocPage = () => (
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     propsTablesMarkdown={[PropsTable]}
     examples={[
+      {
+        example: <TextareaExample />,
+        source: TextareaExampleSrc,
+      },
       {
         example: <FixedSmallSizeExample />,
         source: FixedSmallSizeExampleSrc,