+++ packages/terra-form-checkbox/src/terra-dev-site/doc/form-checkbox/CheckboxField.2.doc.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/CheckboxField.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import CheckboxFieldSrc from '!raw-loader!../../../../src/CheckboxField';
-
 // Examples
+import CheckboxFieldPropsTable from '!terra-props-table-loader!../../../../src/CheckboxField';
+import CheckboxFieldSrc from '!raw-loader!../../../../src/CheckboxField';
 import DefaultCheckboxField from '../example/field/DefaultCheckboxField';
 import DefaultCheckboxFieldSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/field/DefaultCheckboxField';
 import ControlledCheckboxField from '../example/field/ControlledCheckboxField';
@@ -46,8 +46,9 @@ const checkboxFieldExamples = () => (
     ]}
     propsTables={[
       {
-        componentName: 'CheckboxField',
+        componentName: 'Checkbox Field',
         componentSrc: CheckboxFieldSrc,
+        componentProps: CheckboxFieldPropsTable,
       },
     ]}
   />