+++ packages/terra-list/src/terra-dev-site/doc/list/List.1.doc.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import ListSrc from '!raw-loader!../../../../src/List.jsx';
-
 // Example File
+import ListPropsTable from '!terra-props-table-loader!../../../../src/List';
+import ListSrc from '!raw-loader!../../../../src/List';
 import ListExample from '../example/List';
 import ListExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/List';
 import ListDividedExample from '../example/ListDivided';
@@ -41,6 +41,7 @@ const DocPage = () => (
       {
         componentName: 'List',
         componentSrc: ListSrc,
+        componentProps: ListPropsTable,
       },
     ]}
   />