+++ packages/terra-scroll/src/terra-dev-site/doc/scroll/Scroll.1.doc.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import ScrollSrc from '!raw-loader!terra-scroll/src/Scroll';
-
 // Example Files
+import ScrollPropsTable from '!terra-props-table-loader!../../../../src/Scroll';
+import ScrollSrc from '!raw-loader!../../../../src/Scroll';
 import ScrollVertical from '../example/ScrollVertical';
 import ScrollVerticalSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ScrollVertical';
 import ScrollHorizontal from '../example/ScrollHorizontal';
@@ -32,6 +32,7 @@ const DocPage = () => (
       {
         componentName: 'Scroll',
         componentSrc: ScrollSrc,
+        componentProps: ScrollPropsTable,
       },
     ]}
   />