+++ packages/terra-dynamic-grid/src/terra-dev-site/doc/dynamic-grid/DynamicGrid.1.doc.jsx
@@ -3,11 +3,11 @@ import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import DynamicGridSrc from '!raw-loader!../../../../src/DynamicGrid';
-import RegionSrc from '!raw-loader!../../../../src/Region';
-
 // Example Files
+import DynamicGridPropsTable from '!terra-props-table-loader!../../../../src/DynamicGrid';
+import DynamicGridSrc from '!raw-loader!../../../../src/DynamicGrid';
+import RegionPropsTable from '!terra-props-table-loader!../../../../src/Region';
+import RegionSrc from '!raw-loader!../../../../src/Region';
 import OneColumn from '../example/OneColumn';
 import OneColumnSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/OneColumn';
 import TwoColumn from '../example/TwoColumn';
@@ -57,10 +57,12 @@ const DocPage = () => (
       {
         componentName: 'Dynamic Grid',
         componentSrc: DynamicGridSrc,
+        componentProps: DynamicGridPropsTable,
       },
       {
         componentName: 'Region',
         componentSrc: RegionSrc,
+        componentProps: RegionPropsTable,
       },
     ]}
   />