+++ packages/terra-form-radio/src/terra-dev-site/doc/form-radio/RadioField.2.doc.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/RadioField.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import RadioFieldSrc from '!raw-loader!../../../../src/RadioField';
-
 // Examples
+import FormRadioFieldPropsTable from '!terra-props-table-loader!../../../../src/Radio';
+import FormRadioFieldSrc from '!raw-loader!../../../../src/Radio';
 import DefaultRadioField from '../example/field/DefaultRadioField';
 import DefaultRadioFieldSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/field/DefaultRadioField';
 import ControlledRadioField from '../example/field/ControlledRadioField';
@@ -46,8 +46,9 @@ const RadioFieldExamples = () => (
     ]}
     propsTables={[
       {
-        componentName: 'RadioField',
-        componentSrc: RadioFieldSrc,
+        componentName: 'Form Radio Field',
+        componentSrc: FormRadioFieldSrc,
+        componentProps: FormRadioFieldPropsTable,
       },
     ]}
   />