+++ packages/terra-form-select/src/multiple/Frame.jsx
@@ -8,6 +8,7 @@ import Dropdown from '../shared/_Dropdown';
 import Menu from './Menu';
 import FrameUtil from '../shared/_FrameUtil';
 import styles from '../shared/_Frame.module.scss';
+import MenuUtil from '../shared/_MenuUtil';
 
 const cx = classNames.bind(styles);
 
@@ -169,9 +170,9 @@ class Frame extends React.Component {
   }
 
   getDisplay(displayId, ariaDescribedBy) {
-    const { searchValue, isFocused, value } = this.state;
+    const { searchValue, isFocused } = this.state;
     const {
-      disabled, display, placeholder, required,
+      disabled, display, placeholder, required, value,
     } = this.props;
 
     const isHidden = !isFocused && value && value.length > 0;
@@ -321,7 +322,10 @@ class Frame extends React.Component {
    * @param {event} event - The onKeyDown event.
    */
   handleKeyDown(event) {
-    const { value } = this.props;
+    const {
+      children, intl, onDeselect, value,
+    } = this.props;
+
     const { keyCode, target } = event;
 
     if (keyCode === KeyCode.KEY_SPACE && target !== this.input) {
@@ -331,7 +335,17 @@ class Frame extends React.Component {
       event.preventDefault();
       this.openDropdown(event);
     } else if (keyCode === KeyCode.KEY_BACK_SPACE && !this.state.searchValue && value.length > 0) {
-      this.props.onDeselect(value[value.length - 1]);
+      const lastOptionValue = value[value.length - 1];
+      const lastOption = MenuUtil.findByValue(children, lastOptionValue);
+      const lastOptionDisplay = lastOption ? lastOption.props.display : lastOptionValue;
+
+      if (this.visuallyHiddenComponent && this.visuallyHiddenComponent.current) {
+        this.visuallyHiddenComponent.current.innerText = intl.formatMessage({ id: 'Terra.form.select.unselectedText' }, { text: lastOptionDisplay });
+      }
+
+      if (onDeselect) {
+        onDeselect(lastOptionValue);
+      }
     } else if (keyCode === KeyCode.KEY_ESCAPE) {
       this.closeDropdown();
     }