+++ packages/terra-props-table/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+2.40.0 - (September 6, 2019)
+------------------
+### Added
+* Props table markdown file import added to doc site page
+
 2.39.0 - (August 21, 2019)
 --------------------------
 ### Changed